resume_extractor/resume_extractor/requirements.txt
spacy
python-docx
pdfplumber
